# [Performance] Fix for number of threads in COOToCSR