@yaox12 2022-12-29
Perhaps we should update fix the `omp_get_num_threads` in corresponding tests.
https://github.com/dmlc/dgl/blob/025e43210f724ea3a5f308463166e61b4cd97399/tests/cpp/test_spmat_coo.cc#L216
and 
https://github.com/dmlc/dgl/blob/025e43210f724ea3a5f308463166e61b4cd97399/tests/cpp/test_spmat_coo.cc#L293