@anko-intel 2023-01-02
> Perhaps we should fix the `omp_get_num_threads` in corresponding tests.
> 
> https://github.com/dmlc/dgl/blob/025e43210f724ea3a5f308463166e61b4cd97399/tests/cpp/test_spmat_coo.cc#L216
> 
> 
> and
> https://github.com/dmlc/dgl/blob/025e43210f724ea3a5f308463166e61b4cd97399/tests/cpp/test_spmat_coo.cc#L293

I've been thinking about this. 
My guess is that these checks were inserted to ensure that both versions of Unsorted*COOToCSR are tested. But it doesn't work. Fixing the number of threads for the given example could result in a situation where the number of elements/edges is greater than possible (or many edges between the same nodes). For a machine with 64 threads available and the desired value of False from the IsSparse() function, we need to insert at least 64 elements (edges) into a 4x5 matrix!
So the test should be performed on a much larger tensor, which does not seem to be a reasonable and efficient solution. From the point of view of testing, it would be best to have a method to select the COOToCSR version from the API (maybe as an additional "hint" parameter), but from the point of view of the user it seems to be unneceserry complicated. Maybe setting the number of threads from the test would be a workaround.
I'm going to do some measurements and experiments for the COOToCSR function and maybe some other conclusions/improvements will emerge. Let me  wait with the fixes in test until then.